Completed Creation of the functionality of BinarySearchTree as well as EmployeeTreeDecorator
I would say a majority of the functionality if not all of it has been created for everything we need that relates specifically to the EmployeeTreeDecorator.java file and the BinarySearchTree.java files.

We of course still need the Scanner/UI. There might be a couple of tests that need to be rewritten as I have changed the functionality of the compareTo() method within the Employee class to only compare the last names. The EmployeeTreeDecorator.java is very important to meeting our requirements because it offers abstraction of different implementation details away from the user as well as converting Object[] return types from the Generic class BinarySearchTree.java.